+++ CHANGELOG.md
@@ -5,13 +5,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.6.0]
+
+### Added
+
+- plugin state API
+
+### Changed
+
+- title styles
+
 ## [0.5.0]
 
-## Added
+### Added
 
 - the ability to set the settings from the request
 
-## Fixed
+### Fixed
 
 - if there is a title at small panel sizes, the value is out of view
 - if the value is greater than the range, the color of the value is the same as the color of the last range